refactor: Domain 개선

+++ backend/sin/src/main/java/sin/sin/domain/level/Level.java
@@ -21,6 +21,6 @@ public class Level {
     private Long id;
 
     @Column(nullable = false)
-    private float salePercent;
+    private int salePercent;
 
 }
